requirements.txt.txt
python-telegram-bot==20.3
ccxt
pandas
ta
nest_asyncio
flask
